+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 
 # Change Log
 
+### 0.1.13 - ??
+ - Add repository url to repo command output
+
 ### 0.1.12 - Pushing code quality
  - Change the 3 star threshold down to 750 from 1000 recent downloads
  - Add a new sub command repo - for things from github